Release 1: DRY up Teacher and ApprenticeTeacher classes
- Create a subclass of DbcMember called DbcTeacher and make Teacher and
  ApprenticeTeacher classes inherit from it.
- Create another initialize method in DbcTeacher and set the default phase variable to 3 after calling super.
- Pull the #set_phase, #salary=, and #receive_salary methods to DbcTeacher.
- Pull the #set_performance_rating method from both subclasses into DbcTeacher
  and added type checks to if and elsif branches.
