+++ apprentice_teacher.rb
@@ -1,17 +1,10 @@
-require_relative 'dbc_member'
+require_relative 'dbc_teacher'
 
-class ApprenticeTeacher < DbcMember
-  attr_reader :salary, :target_raise
+class ApprenticeTeacher < DbcTeacher
 
   def initialize(options={})
     super
-    @phase = 3
     @target_raise = 800
-  end
-
-  def set_phase(num)
-    @phase = num
-    "Cool, I've always wanted to teach phase #{num}!"
   end
 
   def teach_stuff
@@ -19,26 +12,6 @@ class ApprenticeTeacher < DbcMember
     response += "Listen, class, this is how everything works. "
     response += "*drops crazy knowledge bomb* "
     response += "... You're welcome."
-    response
-  end
-
-  def salary=(new_salary)
-    puts "This better be good!"
-    @salary = new_salary
-  end
-
-  def receive_raise(raise)
-    @salary += raise
-  end
-
-  def set_performance_rating(rating)
-    response = ""
-    if rating > 80
-      response = "Yay, I'm a great employee!"
-      receive_raise(@target_raise)
-    else
-      response = "Oh, well -- thanks to this actionable, specific, and kind feedback, I'll do better next time."
-    end
     response
   end
