+++ teacher.rb
@@ -1,17 +1,11 @@
-require_relative 'dbc_member'
+require_relative 'dbc_teacher'
 
-class Teacher < DbcMember
-  attr_reader :salary, :performance_rating, :target_raise
+class Teacher < DbcTeacher
+  attr_reader :performance_rating
 
   def initialize(options={})
     super
-    @phase = 3
     @target_raise = 1000
-  end
-
-  def set_phase(num)
-    @phase = num
-    "Cool, I've always wanted to teach phase #{num}!"
   end
 
   def teach_stuff
@@ -19,27 +13,6 @@ class Teacher < DbcMember
     response += "Listen, class, this is how everything works, fo SHO! "
     response += "*drops flat-out insane knowledge bomb* "
     response += "... You're welcome. *saunters away*"
-    response
-  end
-
-  def salary=(new_salary)
-    puts "This better be good!"
-    @salary = new_salary
-  end
-
-  def receive_raise(raise)
-    @salary += raise
-  end
-
-  def set_performance_rating(rating)
-    response = ""
-    if rating > 90
-      receive_raise(@target_raise)
-      response = "Yay, I'm a great employee!"
-    else
-      response += "Oh, well -- thanks to this actionable, specific, and kind "
-      response += "feedback, I'll do better next time."
-    end
     response
   end
 end
